Add ExportDataProgress class and refactor DbDataExport to use it
