+++ src/database/dbexportdata.h
@@ -5,32 +5,14 @@
 #include "database.h"
 #include "../threading/cancellation.h"
 
-#include <cstdint>
 #include <utility>
 
-class ExportDataProgress {
-    uint64_t affectedRows = 0;
-    bool isComplete = false;
-
-public:
-    void reset() {
-        affectedRows = 0;
-        isComplete = false;
-    }
-
-    void increment() {
-        affectedRows++;
-        isComplete = false;
-    }
-
-    [[nodiscard]] uint64_t getAffectedRows() const { return affectedRows; }
-    [[nodiscard]] bool isCompleted() const { return isComplete; }
-    void setCompleted() { isComplete = true; }
-};
+#include "progress.h"
 
 class DbDataExport : public DbExport {
 public:
-    explicit DbDataExport(DatabaseInfo info) : DbExport(std::move(info)) {
+    explicit DbDataExport(DatabaseInfo info) :
+        DbExport(std::move(info)) {
     }
 
     void exportDataToFile(const Database *database,
